CRUD operations

+++ demo/src/main/java/com/example/demo/Student.java
@@ -10,7 +10,7 @@ public class Student {
 	
 	 @Id
      @GeneratedValue(strategy = GenerationType.IDENTITY)
-	 private Long id;
+	 private int id;
 	 
     private String name;
     private int age;
@@ -25,11 +25,11 @@ public class Student {
     }
 
     
-    public Long getId()
+    public int getId()
     {
        return id;
     }
-    public void setId(Long id)
+    public void setId(int id)
     { 
     	this.id = id; 
     	}
